[isKingInCheck] unduplicated, moved to helpers
