+++ ui/client/rules/validate-path.js
@@ -2,9 +2,9 @@ module.exports = (origin, destination, matrix) => {
   let clear = true;
 
   const path = [];
-  let [a,b] = origin;
-  let [m,n] = destination;
-  let [x,y] = [Math.sign(m-a), Math.sign(n-b)]; //slope
+  let [a,b] = origin; // { x: 2, y: 3 } --> origin.x, origin.y
+  let [m,n] = destination; //x2,y2
+  let [x,y] = [Math.sign(m-a), Math.sign(n-b)]; //slope //dx,dy
   
   while (!(a === m && b === n )) {
     a += x;
